refactor: cache trans buffer managers
